Move ND dashboard components to core/ directory
