+++ src/core/facilities/AdmissionsVsReleases.js
@@ -26,21 +26,18 @@ import sumBy from "lodash/fp/sumBy";
 import values from "lodash/fp/values";
 
 import { toInteger } from "lodash";
-import {
-  COLORS,
-  COLORS_GOOD_BAD,
-} from "../../../assets/scripts/constants/colors";
-import { configureDownloadButtons } from "../../../utils/downloads/downloads";
-import { filterDatasetByDistrict } from "../../../utils/charts/dataFilters";
+import { COLORS, COLORS_GOOD_BAD } from "../../assets/scripts/constants/colors";
+import { configureDownloadButtons } from "../../utils/downloads/downloads";
+import { filterDatasetByDistrict } from "../../utils/charts/dataFilters";
 import {
   toggleLabel,
   getMonthCountFromMetricPeriodMonthsToggle,
   updateTooltipForMetricType,
-} from "../../../utils/charts/toggles";
-import { sortFilterAndSupplementMostRecentMonths } from "../../../utils/transforms/datasets";
-import { monthNamesWithYearsFromNumbers } from "../../../utils/transforms/months";
-import { METRIC_TYPES } from "../../constants";
-import { metricTypePropType } from "../propTypes";
+} from "../../utils/charts/toggles";
+import { sortFilterAndSupplementMostRecentMonths } from "../../utils/transforms/datasets";
+import { monthNamesWithYearsFromNumbers } from "../../utils/transforms/months";
+import { METRIC_TYPES } from "../../components/constants";
+import { metricTypePropType } from "../../components/charts/propTypes";
 
 const dataCountsMapper = (dataset) => ({
   year: dataset[0].year,
