+++ src/core/programming/FtrReferralsByAge.js
@@ -23,20 +23,20 @@ import pipe from "lodash/fp/pipe";
 import {
   COLORS,
   COLORS_FIVE_VALUES,
-} from "../../../assets/scripts/constants/colors";
-import { configureDownloadButtons } from "../../../utils/downloads/downloads";
+} from "../../assets/scripts/constants/colors";
+import { configureDownloadButtons } from "../../utils/downloads/downloads";
 import {
   filterDatasetBySupervisionType,
   filterDatasetByDistrict,
   filterDatasetByMetricPeriodMonths,
-} from "../../../utils/charts/dataFilters";
+} from "../../utils/charts/dataFilters";
 import {
   tooltipForCountChart,
   tooltipForRateChart,
-} from "../../../utils/charts/tooltips";
-import { toInt } from "../../../utils/transforms/labels";
-import { metricTypePropType } from "../propTypes";
-import { METRIC_TYPES } from "../../constants";
+} from "../../utils/charts/tooltips";
+import { toInt } from "../../utils/transforms/labels";
+import { metricTypePropType } from "../../components/charts/propTypes";
+import { METRIC_TYPES } from "../../components/constants";
 
 const chartId = "ftrReferralsByAge";
 const ageBucketLabels = ["<25", "25-29", "30-34", "35-39", "40<"];
